Added a Project Status to the projects page
Also, I filled in the details of my current Projects.

Also found a bug, where if a duo box has multiple lines of description, it grows bigger than the other duo and causes a layout issue. Would likely take a lot of work to fix, so I'm gonna work around it.