The 404 page wasn't loading the correct CSS file
I'd changed over every page to use style.css now that I can generate custom stylesheets with SASS. Left the 404 page on base.css.

I'd love to change the 404 page over to a template like the rest of the pages are, but with the amount of changes I make, it's not really possible.